:tada: Ranking in progress

+++ app/src/main/java/com/example/howdy/model/XpCharts.kt
@@ -1,8 +1,8 @@
 package com.example.howdy.model
 
 class XpCharts(
-    weekly: List<Int>,
-    monthly: List<Int>
+    weekly: List<Double>,
+    monthly: List<Double>
 ) {
     var weekly = weekly
     var monthly = monthly
